breakup telemetry test
